% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_fun.R
\name{find_fun_queue}
\alias{find_fun_queue}
\title{Find a function}
\usage{
find_fun_queue(FUN, envir, envir_queue)
}
\arguments{
\item{FUN}{A function; as a quoted symbol, a character string or as value.}

\item{envir}{The parent environment}

\item{envir_queue}{The queue environment}
}
\description{
Find functions in various places.  Probably best not to use this...
}

