% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\docType{methods}
\name{pop,queue-method}
\alias{pop,queue-method}
\title{pop}
\usage{
\S4method{pop}{queue}(object)
}
\arguments{
\item{\code{object}}{- An object of class \code{queue}.}
}
\description{
A function used to remove the oldest value from a queue.
}
