% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress_timeout}
\alias{progress_timeout}
\title{Progress bar with timeout}
\usage{
progress_timeout(total, timeout, ..., show = NULL, label = NULL, digits = 0)
}
\arguments{
\item{total}{Total number of expected things.  Use \code{NULL} if
you want to wait on a single thing}

\item{timeout}{The number of seconds to wait}

\item{...}{Additional arguments to
\code{progress::progress_bar$new}}

\item{show}{Flag to indicate if the bar should be displayed.  If
\code{NULL} then the global options \code{queuer.progress_show}
and \code{queuer.progress_suppress} are used to determine if the
bar is shown (with suppress overriding show).  This will
evaluate to \code{TRUE} by default.  A logical flag overrides
\code{queuer.progress_show} but not
\code{queuer.progress_suppress}.}

\item{label}{An optional label to prefix the timeout bar with
(will not be padded with space), or in the case of \code{total =
  NULL} to indicate what is being waited on.}

\item{digits}{The number of digits of accuracy to display the
remaining time in}
}
\description{
Progress bar with timeout
}
