% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\docType{methods}
\name{push,queue-method}
\alias{push,queue-method}
\title{Push}
\usage{
\S4method{push}{queue}(object, val)
}
\arguments{
\item{\code{object}}{- An object of class \code{queue}.}
}
\description{
A function used to push a new value to a queue.
}
