% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{queue}
\alias{queue}
\title{An S4 class to represent a queue.}
\usage{
queue(arr)
}
\value{
An S4 class specifying a queue.
}
\description{
An S4 class to represent a queue.
}
\section{Slots}{

\describe{
\item{\code{arr}}{A numeric \code{vector} indicating the values of the queue. 
Not accessable to the user.}

\item{\code{size}}{A number indicating the number of elements in the queue.}
}}

