% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_bundle.R
\name{task_bundle}
\alias{task_bundle}
\title{Create a task bundle}
\usage{
task_bundle(obj, task_ids = NULL, group = NULL, names = NULL)
}
\arguments{
\item{obj}{An observer or queue object; something that can be
passed through to \code{\link{context_db}}.}

\item{group}{Group name}

\item{names}{Optional vector of names to label output with.}

\item{tasks}{A list of tasks.}
}
\description{
Create a task bundle.  Generally these are not created manually,
but this page serves to document what task bundles are and the
methods that they have.
}
\details{
A task bundle exists to group together tasks that are related.  It
is possible for a task to belong to multiple bundles.
}

