% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_bundle.R
\name{task_bundle_combine}
\alias{task_bundle_combine}
\title{Combine task bundles}
\usage{
task_bundle_combine(..., bundles = list(...), name = NULL,
  overwrite = FALSE)
}
\arguments{
\item{...}{Any number of task bundles}

\item{bundles}{A list of bundles (used in place of \code{...} and
probably more useful for programming).}

\item{name}{Group name}

\item{overwrite}{Logical indicating if an existing bundle with the
same name should be overwritten.  If \code{FALSE} and a bundle
with this name already exists, an error will be thrown.}
}
\description{
Combine two or more task bundles
}
\details{
For now task bundles must have the same function to be combined.
}

