% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{time_checker}
\alias{time_checker}
\title{Simple timer}
\usage{
time_checker(timeout, remaining = FALSE)
}
\arguments{
\item{timeout}{The length of time to set the timer for (can be
infinite)}

\item{remaining}{Logical, indicating if the timer should return
the time remaining (as the number of seconds).  If \code{FALSE}
(the default) then the timer returns \code{FALSE} until the time
has completed, then returns \code{TRUE}.}
}
\description{
A simple timer
}

