% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_bundle.R
\name{task_bundle_create}
\alias{task_bundle_create}
\alias{task_bundle_get}
\title{Create a task bundle}
\usage{
task_bundle_create(task_ids, obj, name = NULL, X = NULL,
  overwrite = FALSE, homogeneous = NULL)

task_bundle_get(name, root)
}
\arguments{
\item{task_ids}{A vector of task ids}

\item{obj}{An observer or queue object; something that can be
passed through to \code{\link{context_root_get}}.}

\item{name}{Group name}

\item{X}{Metadata to associate with the tasks.  This is used by
the bulk interface (\code{\link{qlapply}} and
\code{\link{enqueue_bulk}} to associate the first argument with
the bundle).}

\item{overwrite}{Logical indicating if an existing bundle with the
same name should be overwritten.  If \code{FALSE} and a bundle
with this name already exists, an error will be thrown.}
}
\description{
Create a task bundle.  Generally these are not created manually,
but this page serves to document what task bundles are and the
methods that they have.
}
\details{
A task bundle exists to group together tasks that are related.  It
is possible for a task to belong to multiple bundles.
}

